% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runACF.R
\name{acffunc}
\alias{acffunc}
\title{calculate correlation for residuals by block}
\usage{
acffunc(block, model, suppress.printout = FALSE)
}
\arguments{
\item{block}{Vector of blocks that identify data points that are correlated}

\item{model}{Fitted model object (glm or gam)}
}
\description{
calculate correlation for residuals by block
}

