% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/AICh.R
\name{AICh}
\alias{AICh}
\title{Function to calculate AICh (Hardin and Hilbe 2013)}
\usage{
AICh(model)
}
\description{
Function to calculate AICh (Hardin and Hilbe 2013)
}
\author{
Lindesay Scott-Hayward, University of St Andrews
}

