% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anova.gammrsea.R
\name{anova.gamMRSea}
\alias{anova.gamMRSea}
\title{Anova Tables for \code{gamMRSea} Models}
\usage{
\method{anova}{gamMRSea}(object, varshortnames = NULL, panelid = NULL,
  test = "Wald")
}
\arguments{
\item{object}{A \code{gamMRSea} model object}

\item{varshortnames}{(default = NULL).  Character vector denoting the short names to use for any smooth terms.  May already be specified as part of the model object.}

\item{panelid}{vector of length of the data used in object.  Specified if robust standard errors are to be used.}

\item{test}{(default='wald'). May also specify "F".}
}
\value{
An object of class "\code{anova}".
}
\description{
Calculates type-III analysis-of-variance tables for model objects produced by gamMRSea (in the MRSea package). Wald chisquare tests are calculated by default although, F-tests may be specified.
}
\examples{

# load data
data(ns.data.re)
ns.data.re$foldid<-getCVids(ns.data.re, folds=5)
 
model<-gamMRSea(birds ~ observationhour + as.factor(floodebb) + as.factor(impact),  
              family='poisson', data=ns.data.re)
anova(model)        
             
}

