% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.orig.data.R
\name{bootstrap.orig.data}
\alias{bootstrap.orig.data}
\title{Obtaining a data frame of bootstrapped data using resamples}
\usage{
bootstrap.orig.data(orig.data, resample, new.resamples, resamples.no)
}
\arguments{
\item{orig.data}{Original data to be bootstrapped}

\item{resample}{Specifies the resampling unit for bootstrapping, default is \code{transect.id}. Must match a column name in \code{orig.data} exactly}

\item{new.resamples}{String of resampled units from \code{data[,"resample"]}. Created by \code{create.bootstrap.data()}}

\item{resamples.no}{Length of new.resamples}
}
\value{
Returns bootstrapped data. Internal function called by function \code{create.bootstrap.data}.
}
\description{
This function extracts the records corresponding to each resample from the original 
distance data and pastes them together in a new data frame which is returned.
}
\examples{
data(dis.data.re)
resample<-"transect.id"
samples<-unique(dis.data.re[,resample])
resamples.no<-length(samples)
new.resamples<-sample(samples,resamples.no,replace=TRUE)
bootstrap.data<-bootstrap.orig.data(dis.data.re,resample,new.resamples,resamples.no)

}

