% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/choose.radii.R
\name{choose.radii}
\alias{choose.radii}
\title{Function to choose the radii for the CReSS local radial basis function}
\usage{
choose.radii(
  currentFit,
  indices,
  radiusIndices,
  radii,
  out.lm,
  dists,
  aR,
  baseModel,
  fitnessMeasure,
  response,
  models,
  interactionTerm,
  data,
  initDisp,
  fit.opts,
  basis,
  printout
)
}
\description{
Function to choose the radii for the CReSS local radial basis function
}
