% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drop.step_2d.R
\name{drop.step_2d}
\alias{drop.step_2d}
\title{Function that tries dropping knots to find an improvement in fit}
\usage{
drop.step_2d(
  radii,
  invInd,
  dists,
  explData,
  response,
  knotgrid,
  maxIterations,
  fitnessMeasure,
  point,
  knotPoint,
  position,
  aR,
  BIC,
  track,
  out.lm,
  improveDrop,
  minKnots,
  tol = 0,
  baseModel,
  radiusIndices,
  models,
  interactionTerm,
  data,
  initDisp,
  fit.opts,
  basis,
  printout
)
}
\description{
Function that tries dropping knots to find an improvement in fit
}
\author{
Cameron Walker, Department of Engineering Science, University of Auckland.
}
