% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.thinplate_2d.R
\name{fit.thinPlate_2d}
\alias{fit.thinPlate_2d}
\title{Function to fit a local radial basis function (CReSS) as a two dimensional smooth}
\usage{
fit.thinPlate_2d(
  fitnessMeasure,
  dists,
  aR,
  radii,
  baseModel,
  radiusIndices,
  models,
  currentFit,
  interactionTerm,
  data,
  initDisp,
  fit.opts,
  basis = "gaussian",
  printout
)
}
\description{
Function to fit a local radial basis function (CReSS) as a two dimensional smooth
}
