% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamMRSea.R
\name{gamMRSea}
\alias{gamMRSea}
\title{gamMRSea model function}
\usage{
gamMRSea(formula, family = gaussian, data, weights, subset, na.action,
  start = NULL, etastart, mustart, offset, control = list(...),
  model = TRUE, method = "glm.fit", x = FALSE, y = TRUE,
  contrasts = NULL, splineParams = NULL, ...)
}
\description{
gamMRSea model function
}

