% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCV_cress.R
\name{getCV_CReSS}
\alias{getCV_CReSS}
\title{Calculate cross-validation score for a CReSS type model}
\usage{
getCV_CReSS(datain, baseModel, splineParams = NULL, vector = FALSE)
}
\arguments{
\item{datain}{Data frame containing columns of covariates contained in \code{baseModel}.}

\item{baseModel}{'glm' or 'gamMRSea' model object}

\item{splineParams}{list object containing information for fitting one and two dimensional splines. See \code{\link{makesplineParams}} for more details.}

\item{vector}{Logical indicating whether the vector of scores is returned (TRUE) or if the mean score is returned (FALSE).}
}
\description{
Calculate cross-validation score for a CReSS type model
}
\details{
There must be a column in the data called \code{foldid}, which can be created using \code{\link{getCVids}}.  This column defines the folds of data for the CV calculation.  If this is not provided, by default random allocation is given to 5 folds.

The cost function for this CV is a mean squared error.
}
\examples{
# load data
data(ns.data.re)
ns.data.re$foldid<-getCVids(ns.data.re, folds=5)
 
model<-gamMRSea(birds ~ observationhour + as.factor(floodebb) + as.factor(impact),  
              family='poisson', data=ns.data.re)
              
# calculate CV
getCV_CReSS(ns.data.re, model)

}
\author{
LAS Scott-Hayward (University of St Andrews)
}

