% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{getCVpp}
\alias{getCVpp}
\title{Get block leave-one-out CV score for point process model}
\usage{
getCVpp(model.obj, folddata, kfold = 10, replicates = 1)
}
\arguments{
\item{model.obj}{a point process gamMRSea model}

\item{replicates}{(default = 100). number stating how many times to replicate the CV score}

\item{proj4string}{string giving CRS of x.pos and y.pos in date in \code{model.obj}}

\item{blocksize}{size of the blocks, give in m if \code{proj4string} is in degrees or m, or in km if data in  \code{model.obj}/\code{proj4string} is in km}

\item{folds}{(default = 10). number stating many folds for CV}

\item{showProgress}{(default = FALSE) logical stating whether to print progress bar}
}
\description{
This function calculates a leave-one-out cross-validation score for point process models
}
