% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDispersion.R
\name{getDispersion}
\alias{getDispersion}
\title{dispersion parameter}
\usage{
getDispersion(model)
}
\arguments{
\item{model}{gamMRSea or glm model object}
}
\value{
single number of dispersion parameter estimation
}
\description{
This function calculates the dispersion parameter for Normal, Binomial, Poisson and Gamma distributions
}
\details{
some details
}
\author{
LAS Scott-Hayward
}
