% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getEmpDistribution.R
\name{getEmpDistribution}
\alias{getEmpDistribution}
\title{Function to generate the empirical distribution of the runs test statistic, given some data and a model.}
\usage{
getEmpDistribution(n.sim, simData, model, data, plot = FALSE,
  returnDist = TRUE, dots = TRUE)
}
\arguments{
\item{n.sim}{The number of simulated sets of data.}

\item{simData}{A matrix, where each column is a set of data simulated under independence, with rows the same length as the data used for the model.}

\item{model}{a glm or gamMRSea model object}

\item{data}{data set used to fit the model.}

\item{plot}{logical flag. If TRUE, a plot is made showing the 5% critical values for the empirical distribution vs the N(0,1) distribution. Default is `FALSE`}

\item{returnDist}{logical flag.  Do you want the distribution of test statistics returned ('TRUE') or just the 5% critical values ('FALSE')}

\item{dots}{(Default: \code{TRUE}. Logical stating whether to show a printout of block numebers to assess progress. `TRUE` will print dots into the workspace.}
}
\description{
Function to generate the empirical distribution of the runs test statistic, given some data and a model.
}
\examples{
data(ns.data.re)

model<-gamMRSea(birds ~ observationhour + as.factor(floodebb) + as.factor(impact),  
              family='poisson', data=ns.data.re)

simData<-generateNoise(n=500, response=fitted(model), family='poisson')
empdist<-getEmpDistribution(500, simData, model, data=ns.data.re, plot=FALSE, returnDist=TRUE,dots=FALSE)

}
\author{
LAS Scott-Hayward, University of St Andrews
}

