% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getKnotgrid.R
\name{getKnotgrid}
\alias{getKnotgrid}
\title{Generate a grid of knot locations to run SALSA2D.}
\usage{
getKnotgrid(coordData, numKnots = 300, plot = TRUE)
}
\arguments{
\item{coordData}{nx2 matrix or data frame of coordinates representing data locations}

\item{numKnots}{(\code{default = 300)}).  The user may choose how many legal knot locations are available (only if more than 400 )}

\item{plot}{(\code{default = TRUE}). Logical stating whether a plot showing the legal knot positions is given.}
}
\value{
A (10000 x 2) matrix of knot locations.  Any row with an illegal knot location (e.g. not near data) contains c(NA, NA)
}
\description{
Generate a grid of knot locations to run SALSA2D.
}
\details{
SALSA2D requires a grid of knot locations to determine the best locations.  Illegal knot positions (those not close to data) are kept as a row in the data frame of locations but given c(NA, NA) to avoid a knot considered.
}
\examples{
 \dontrun{
data(dis.data.re)
# bootstrap data without stratification
dis.data.re$survey.id<-paste(dis.data.re$season,dis.data.re$impact,sep="")
require(mrds)
result<-ddf(dsmodel=~mcds(key="hn", formula=~1), data=dis.data.re, method="ds",
             meta.data=list(width=250))
dis.data.re<-create.NHAT(dis.data.re,result)

knotgrid<-getKnotgrid(cbind(dis.data.re$x.pos, dis.data.re$y.pos))}

}
