% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{getlargestresid}
\alias{getlargestresid}
\title{Get index of n largest residuals}
\usage{
getlargestresid(model, n = 5)
}
\arguments{
\item{model}{a gamMRSea model}

\item{n}{number of largest residuals to return}
}
\description{
This function evaluates the 5 largest residuals from a model.
}
