% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{getPPresiduals}
\alias{getPPresiduals}
\title{Get Pearsons residuals for point process model}
\usage{
getPPresiduals(model, type = "Pearson")
}
\arguments{
\item{model}{A point process model fitted in a glm framework}

\item{type}{Type of residual required. Choices are "Pearson" (default) or "raw"}
}
\description{
This function calculates the residuals for a point process model.
}
