% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getRadiiSequence.R
\name{getRadiiSequence}
\alias{getRadiiSequence}
\title{Function for obtaining a sequence of range parameters for the bivariate CReSS smoother}
\usage{
getRadiiSequence(
  method = NULL,
  numberofradii = 10,
  distMatrix,
  basis,
  rvin = NULL,
  xydata,
  response,
  alpha = 0,
  vgmmodel = "Sph",
  showplots = FALSE,
  ...
)
}
\arguments{
\item{method}{One of "original" or "variogram". The original method is described in Scott-Hayward et al 2013 (reference below) and the variogram method is in the \href{https://lindesaysh.github.io/MRSea/articles/web/UserRadiiChoice_MRSea.html}{vignette for variogram radii selection}}

\item{numberofradii}{The number of range parameters for SALSA to use when fitting the CReSS smooth.  The default is 8.  Remember, the more parameters the longer SALSA will take to find a suitable one for each knot location.}

\item{distMatrix}{Matrix of distances between data locations and knot locations (n x k). May be Euclidean or geodesic distances. Euclidean distances created using \code{\link{makeDists}}.}

\item{basis}{character stating whether a 'gaussian' or 'exponential' basis is being used.
#' @param rvin Optional if \code{method = "original"}. Two parameter vector stating the minimum and maximum range of r for a gaussian basis.}

\item{xydata}{Required if `method = "variogram". Data frame containing columns for x and y coordinates. x is assumed to be the first of the two columns}

\item{response}{Required if \code{method = "variogram"}. Vector of response values for use in \code{gstat::variogram}.  These values should be approximately normally distributed.}

\item{alpha}{Optional if \code{method = "variogram"}. Numeric parameter for the \code{gstat::variogram} function giving the direction in plane(x,y)}

\item{showplots}{(\code{default = FALSE}). Optional if \code{method = "variogram"}. If \code{TRUE} the output of \code{gstat::variogram} and \code{gstat::fit.variogram} are shown.}

\item{...}{Other parameters for the \code{gstat::variogram} function.}
}
\value{
\itemize{
\item Original method: This function returns a vector containing a sequence of range parameters.
\item Variogram method: This function returns a vector containing a sequence of range parameters.  If an even number of radii is requested, this is reduced by one to give an odd length sequence where the middle number was the best range parameter from the variogram. The outputs of the variogram model can be found in the attributes of the returned object under \code{vg.fit}.
}

\strong{Note} If the estimated range from the variogram model exceeds the maximum distance in the distance matrix, the sequence reverts to the original method. A warning will be printed and the method used is an attribute in the returned sequence ("Method").
}
\description{
Function for obtaining a sequence of range parameters for the bivariate CReSS smoother
}
\details{
The range parameter determines the range of the influence of each knot.  Small numbers indicate local influence and large ones, global influence.
}
\examples{

# load data
data(ns.data.re)
rad.dat <- dplyr::filter(ns.data.re, impact==0, Year==9, MonthOfYear == 3)
# load knot grid data
data(knotgrid.ns)

# make distance matrices for datatoknots and knottoknots
distMats<-makeDists(cbind(rad.dat$x.pos, rad.dat$y.pos), na.omit(knotgrid.ns))
# choose sequence of radii
r_seq<-getRadiiSequence(method = "variogram",
                        numberofradii = 8, 
                        xydata = rad.dat[,c("x.pos", "y.pos")], 
                        response = log(rad.dat$birds +1 ), 
                        basis = "gaussian", 
                        distMatrix = distMats$dataDist)

r_seq
attr(r_seq, "vg.fit")
attr(r_seq, "Method")

 r_seq<-getRadiiSequence(method = "original",
                         numberofradii = 8, 
                         distMatrix = distMats$dataDist, 
                         basis="gaussian")

r_seq
attr(r_seq, "Method")

}
\references{
Scott-Hayward, L.; M. Mackenzie, C.Donovan, C.Walker and E.Ashe.  Complex Region Spatial Smoother (CReSS). Journal of computational and Graphical Statistics. 2013.
DOI: 10.1080/10618600.2012.762920

\href{https://lindesaysh.github.io/MRSea/articles/web/UserRadiiChoice_MRSea.html}{vignette for variogram radii selection}
}
