% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LRF.g.R
\name{LRF.g}
\alias{LRF.g}
\title{Function for creating an Gaussian basis function for a spatial smooth using the CReSS method.}
\usage{
LRF.g(radiusIndices, dists, radii, aR)
}
\arguments{
\item{radiusIndices}{Vector of length startKnots identifying which radii (\code{splineParams[[1]]$radii}) will be used to initialise the model}

\item{dists}{Matrix of distances between data locations and knot locations (n x k). May be Euclidean or geodesic distances.}

\item{radii}{Sequence of range parameters for the CReSS basis from local (small) to global (large). Determines the range of the influence of each knot.}

\item{aR}{Index of knot locations. The index contains numbers selected by SALSA from 1 to the number of legal knot locations \code{na.omit(knotgrid)}. Used to specify which columns of \code{dists} should be used to construct the basis matrix.}
}
\value{
Returns a basis matrix with one column for each knot in \code{aR} and one row for every observation (i.e. same number of rows as \code{dists})
}
\description{
This function calculates a local radial Gausiian basis matrix for use in \code{\link{runSALSA2D}}.
}
\details{
Calculate a local radial basis matrix for use in \code{\link{runSALSA2D}}.  The distance matrix input may be Euclidean or geodesic distances.
}
\examples{

# load data
data(ns.data.re)
# load knot grid data
data(knotgrid.ns)

splineParams<-makesplineParams(data=ns.data.re, varlist=c('observationhour'))

# make distance matrices for datatoknots and knottoknots
distMats<-makeDists(cbind(ns.data.re$x.pos, ns.data.re$y.pos), na.omit(knotgrid.ns), knotmat=FALSE)

# choose sequence of radii
r_seq<-getRadiiSequence(method = "variogram",
                        numberofradii = 8, 
                        xydata = ns.data.re[,c("x.pos", "y.pos")], 
                        response = log(ns.data.re$birds +1 ), 
                        basis = "gaussian", 
                        distMatrix = distMats$dataDist)

# using the fourth radius and picking 5 knots
basis<-LRF.g(radiusIndices=rep(4, 5), dists=distMats$dataDist, radii = r_seq, 
        aR=c(3, 10, 15, 28, 31))

}
