% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make.gamMRSea.R
\name{make.gamMRSea}
\alias{make.gamMRSea}
\title{Function to make model of class \code{gamMRSea}}
\usage{
make.gamMRSea(model, panelid = NULL, splineParams = NULL,
  varshortnames = NULL, gamMRSea = FALSE)
}
\arguments{
\item{model}{model object of class glm or gamMRSea}

\item{panelid}{vector of length of the data containing the panel identification for each row of data}

\item{splineParams}{MRSea based list object}

\item{varshortnames}{vector containing the short names for each variable.  These are used in summary and anova}

\item{gamMRSea}{logical stating whether the call of the model should be changed to 'gamMRSea' from `glm`}
}
\description{
Function to allow a model of class \code{gamMRSea} to be updated to include a panel structure or shortnames to make summary and anova outputs more readable. Function to update an `lm` or `glm` model to class `gamMRSea`.
}
\author{
LAS Scott-Hayward, University of St Andrews
}

