% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{makeDists}
\alias{makeDists}
\title{Make Euclidean distance matrices for use in CReSS and SALSA model frameworks}
\usage{
makeDists(
  datacoords,
  knotcoords,
  knotmat = TRUE,
  polys = NULL,
  type = "A",
  plot.transition = FALSE,
  grid.dim = c(100, 100)
)
}
\arguments{
\item{datacoords}{Coordinates of the data locations. If a design matrix interaction is required, add a third column containing the factor level.}

\item{knotcoords}{Coordinates of the legal knot locations. If a design matrix interaction is required, add a third column containing the factor level for each knot location.}

\item{knotmat}{(\code{default=TRUE}). Should a matrix of knot-knot distances be created}

\item{polys}{(\code{default=NULL}). If geodesic distances are to be calculated, provide a list of polygons defining exclusion areas.}

\item{type}{(\code{default='A'}).  One of 'A' or 'B'.  'A' is used when the \code{knotcoords} are a subset of \code{datacoords} AND the attributes of \code{knotcoords} give the index of points from \code{datacoords} (as happens if \code{getKnotgrid()} is used).  'B' is used for prediction (when \code{datacoords} is a prediction grid and so \code{knotcoords} is not a subset) or when the knotgrid was not generated using \code{getKnotgrid()}.}

\item{plot.transition}{(\code{default=TRUE}). Logical stating whether to plot the transition matrix.  Useful to see if the boundaries are being obeyed.}

\item{grid.dim}{This is a vector of length two which specifies the dimesions of the grid used to create the transition matrix (default \code{c(100, 100)}.  If the transition matrix shows that boundaries are being ignored, the grid dimensions will need to increase. However, increasing the grid, whilst improving accuracy, also increases computational time.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

This function makes two Euclidean distance matrices.  One for the distances between all spatial observations and all spatial knot locations.  The other, if specified, is the distances between knot locations. It is experimental owing to the addition of the creation of infinity-block distance matrices. The original functionality is unchanged.
}
\details{
The data-knot matrix is used in the CReSS basis and the knot-knot matrix is used in SALSA to determine where a nearest knot to `move' should be.

If three columns are provided for \code{datacoords} and \code{knotcoords} the matrix returned has infinity for distances between knots and data associated with differing factor levels.
}
\examples{
# load data
data(ns.data.re)
# load knot grid data
data(knotgrid.ns)
 
# make distance matrices for datatoknots and knottoknots
distMats<-makeDists(cbind(ns.data.re$x.pos, ns.data.re$y.pos), na.omit(knotgrid.ns))

# ~~~~~~~~~~~~~~~~~~
# Example with block-infinity distance matrix
data(nysted.analysisdata)
myknots <- selectFctrKnots(nysted.analysisdata[,c('x.pos', 'y.pos', 'impact')], nk=150)

dists <- makeDists(datacoords = nysted.analysisdata[,c('x.pos', 'y.pos', 'impact')], 
                   knotcoords = myknots, 
                   knotmat = TRUE)
}
