% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/makesplineParams.R
\name{makesplineParams}
\alias{makesplineParams}
\title{Constructing an object of spline parameters}
\usage{
makesplineParams(data, varlist, predictionData = NULL, degree = NULL)
}
\arguments{
\item{data}{Data frame containing columns of covariates listed in \code{varlist}.  Column names must match with names in \code{varlist}}

\item{varlist}{Vector of variable names for the covariates of interest}

\item{predictionData}{Data frame containing columns of covariates listed in \code{varlist}.  Column names must match with those in \code{varlist}.  This parameter is used to find the maximum range of covariates between the data and prediction data. If \code{predictionData} is \code{NULL} then the range of the data is used.}

\item{degree}{Vector specifying the degree of the spline. If unspecified, degree 2 is stored.}
}
\description{
This function makes a list object containing all of the information to fit splines to continuous data.
}
\details{
The information is stored in list slots \code{[[2]]} and onward (slot \code{[[1]]} is reserved for a spatial term). Specifically:

\code{covar}. Name of covariate.

\code{explanatory}. Vector of covariate data.

\code{knots}. Knot(s) for spline fitting.  This function initialises with a knot at the mean covariate value.

\code{bd}. This specifies the boundary knots.  If \code{predictionData} is \code{NULL} then this is the range of the covariate data.  Otherwise, the boundary knots are the maximum combined range of the data and prediction data.

\code{degree}. The degree of a B-spline. This function retuns 2 by default.


See \code{\link{runSALSA2D}} for details on the spatial slot (\code{[[1]]})
}
\examples{
# load data
data(ns.data.re)
# load prediction data
data(ns.predict.data.re)

splineParams<- makesplineParams(ns.data.re, varlist=c('observationhour', 'DayOfMonth'),
                predictionData=ns.predict.data.re)
}

