% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.frame.gamMRSea.R
\name{model.frame.gamMRSea}
\alias{model.frame.gamMRSea}
\title{Calculation of model fram for gamMRSea model object}
\usage{
\method{model.frame}{gamMRSea}(
  formula,
  data = NULL,
  subset = NULL,
  na.action = na.fail,
  drop.unused.levels = FALSE,
  xlev = NULL,
  splineParams = NULL,
  ...
)
}
\description{
Calculation of model fram for gamMRSea model object
}
\note{
Code adapted from \code{model.frame.glm}
}
\author{
Lindesay Scott-Hayward, University of St Andrews.
}
