% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadescriptions_nysted.R
\docType{data}
\name{nysted.analysisdata}
\alias{nysted.analysisdata}
\title{Nysted Data}
\format{
A data frame with 9232 rows and 12 variables
}
\description{
A simulated dataset containing the distance corrected `dis.data.re`. The variables are as follows:
}
\details{
\itemize{
\item \code{transect.id} Identifier for the individual visits to the transects
\item \code{transect.label} Labels for transects
\item \code{season} Numerical indicator for the four different seasons
\item \code{impact} Numerical indicator for before (0) and after (1) impact
\item \code{segment.id} Identifier for individual visits to the segment
\item \code{segment.label} Label for segments
\item \code{length} Length of segment in km
\item \code{x.pos} spatial location in the horizontal axis in UTMs
\item \code{y.pos} spatial location in the vertical axis in UTMs
\item \code{depth} Depth in m
\item \code{area} area of each segment in kmsq
\item \code{NHAT} Bird counts in each segment
\item \code{response} Same as for `NHAT` 
}
}
\keyword{datasets}
