% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadescriptions_nysted.R
\docType{data}
\name{nysted.coast}
\alias{nysted.coast}
\title{Polygon of the Nysted coastline}
\format{
A data frame with 46 rows and 2 variables
}
\description{
The variables are as follows:
}
\details{
\itemize{
\item \code{x} spatial location in the horizontal axis in UTMs. Unit is kilometers.
\item \code{y} spatial location in the vertical axis in UTMs. Unit is kilometers.
}
}
\keyword{datasets}
