% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadescriptions_nysted.R
\docType{data}
\name{nysted.predictdata}
\alias{nysted.predictdata}
\title{Prediction grid data for Nysted Data}
\format{
A data frame with 37928 rows and 10 variables
}
\description{
A simulated prediction dataset containing the true counts, the effort data and other variables of
grid data. The variables are as follows:
}
\details{
\itemize{
\item \code{area} area surveyed in the gridcell in km squared
\item \code{x.pos} spatial location in the horizontal axis in UTMs (km)
\item \code{y.pos} spatial location in the vertical axis in UTMs (km)
\item \code{depth} depth in m
\item \code{segment.id} Identifier for individual visits to the segment
\item \code{season} Numerical indicator for the four different seasons
\item \code{impact} Numerical indicator for before (0) and after (1) impact
\item \code{truth.re} number of birds in the redistribution simulation
\item \code{truth.de} number of birds in the site-wide decrease simulation
\item \code{truth.no} number of birds in the no change simulation
}
}
\keyword{datasets}
