% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadescriptions_nysted.R
\docType{data}
\name{nysted.studybnd}
\alias{nysted.studybnd}
\title{Polygon of the outline of the Nysted study area}
\format{
A data frame with 2210 rows and 2 variables
}
\description{
The variables are as follows:
}
\details{
\itemize{
\item \code{x.pos} spatial location in the horizontal axis in UTMs. Unit is meters.
\item \code{y.pos} spatial location in the vertical axis in UTMs. Unit is meters.
}
}
\keyword{datasets}
