% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCumRes.R
\name{plotCumRes}
\alias{plotCumRes}
\title{Calculate cumulative residuals and plot.}
\usage{
plotCumRes(model, varlist, d2k = NULL, splineParams = NULL, label = "",
  save = FALSE)
}
\arguments{
\item{model}{Fitted model object (glm or gam)}

\item{varlist}{Vector of covariate names (continous covariates only)}

\item{d2k}{(default=NULL).  Distance matrix of data to knot points. Used only if there is a \code{\link{LocalRadialFunction}} smooth in the model formula}

\item{splineParams}{(default \code{=NULL}) List object containing output from runSALSA/runSALSA2D required for updating \code{model}.  Used only if there is a \code{LocalRadialFunction} smooth in the model formula. See \code{\link{makesplineParams}} for details of this object.}

\item{label}{Label printed at the end of the plot name to identify it if \code{save=TRUE}.}

\item{save}{(\code{default=FALSE}). Logical stating whether plot should be saved into working directory.}
}
\value{
Cumulative residual plots are returned for residuals ordered by each covariate in \code{varlist}, predicted value and index of observations (temporally).
The blue dots are the residuals
The black line is the line of cumulative residual.
On the covariate plots (those in \code{varlist}) the grey line indicates what we would expect from a well fitted covariate. i.e. one that is fitted with excessive knots.

Note: if the covariate is discrete in nature (like the example below), there will be a lot of overplotting of residuals.
}
\description{
The output is plots of cumulative residuals.
}
\examples{
# load data
data(ns.data.re)

model<-gamMRSea(birds ~ observationhour + as.factor(floodebb) + as.factor(impact), 
           family='quasipoisson', data=ns.data.re)

plotCumRes(model, varlist=c('observationhour'))

}
