% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadescriptions.R
\docType{data}
\name{predict.data.de}
\alias{predict.data.de}
\title{Prediction grid data for post-impact decrease}
\format{
A data frame with 37928 rows and 8 variables
}
\description{
A simulated dataset containing the true number of birds, the effort data and other variables of 
prediction grid data. The variables are as follows:
}
\details{
\itemize{
\item \code{area} area surveyed in the gridcell in km squared
\item \code{x.pos} spatial location in the horizontal axis in UTMs
\item \code{y.pos} spatial location in the vertical axis in UTMs
\item \code{depth} depth in m
\item \code{segment.id} Identifier for individual visits to the segment
\item \code{season} Numerical indicator for the four different seasons
\item \code{impact} Numerical indicator for before (0) and after (1) impact
\item \code{truth} number of birds
}
}
\keyword{datasets}
