% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{print.gamMRSea}
\alias{print.gamMRSea}
\title{Printing of gamMRSea model object}
\usage{
\method{print}{gamMRSea}(x, digits = max(3L, getOption("digits") - 3L), ...)
}
\description{
Printing of gamMRSea model object
}
\note{
Code adapted from \code{print.glm}
}
\author{
Lindesay Scott-Hayward, University of St Andrews.
}
