% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{print.summary.gamMRSea}
\alias{print.summary.gamMRSea}
\title{Printing of summary of gamMRSea model object}
\usage{
\method{print}{summary.gamMRSea}(
  x,
  digits = max(3L, getOption("digits") - 3L),
  symbolic.cor = x$symbolic.cor,
  signif.stars = getOption("show.signif.stars"),
  ...
)
}
\description{
Printing of summary of gamMRSea model object
}
\note{
Code adapted from \code{print.summary.glm}
}
\author{
Lindesay Scott-Hayward, University of St Andrews.
}
