% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateNoise.R
\name{qzibinom}
\alias{qzibinom}
\title{qzibinom function from the VGAM package}
\usage{
qzibinom(p, size, prob, pstr0 = 0)
}
\description{
qzibinom function from the VGAM package
}
\author{
VGAM package
}

