% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runACF.R
\name{runACF}
\alias{runACF}
\title{run functions to create acf matrix and plot the results}
\usage{
runACF(
  block,
  model,
  store = FALSE,
  save = F,
  suppress.printout = TRUE,
  maxlag = NULL,
  printplot = TRUE
)
}
\arguments{
\item{block}{Vector of blocks that identify data points that are correlated}

\item{model}{Fitted model object (glm or gam)}

\item{store}{(\code{default=FALSE}). Logical stating whether a list of the matrix of correlations is stored (output from \code{acffunc}.)}

\item{save}{(\code{default=FALSE}). Logical stating whether plot should be saved into working directory.}

\item{suppress.printout}{(\code{default=TRUE}. Logical stating whether to show a printout of block numbers to assess progress. \code{FALSE} will show printout.}

\item{maxlag}{(\code{default=NULL}). Numeric entry to allow the restriction of the maximum lag on the plots.  If \code{NULL} then the length of the longest panel is used as the maximum plotted lag.}
}
\value{
Plot of lag vs correlation.  Each grey line is the correlation for each individual block in \code{block}.  The red line is the mean values for each lag.

If \code{store=TRUE} then the matrix of correlations (nblocks x length_max_block) is returned and \code{plotacf} may be used to plot the acf.
}
\description{
run functions to create acf matrix and plot the results
}
\examples{
# load data
data(ns.data.re)

model<-gamMRSea(birds ~ observationhour + as.factor(floodebb) + as.factor(impact), 
           family='quasipoisson', data=ns.data.re)

ns.data.re$blockid<-paste(ns.data.re$GridCode, ns.data.re$Year, ns.data.re$MonthOfYear, 
                    ns.data.re$DayOfMonth, sep='')
ns.data.re$blockid<-as.factor(ns.data.re$blockid)

runACF(ns.data.re$blockid, model, suppress.printout=TRUE)

# storing the output and then plotting
acfoutput <- runACF(ns.data.re$blockid, model, suppress.printout=TRUE, store=TRUE)
plotacf(acfoutput$acfmat)

}
\author{
LAS Scott-Hayward, University of St Andrews
}
