% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runInfluence.R
\name{runInfluence}
\alias{runInfluence}
\title{Assessing the influece of each correlated block on both the precision of the parameter estimates (COVRATIO statistics) and the sensitivity of model predictions (PRESS statistics).}
\usage{
runInfluence(model, id = NULL, save = FALSE, dots = FALSE)
}
\arguments{
\item{model}{Fitted model object (glm, gamMRSea or gam)}

\item{id}{blocking structure}

\item{save}{(\code{default=FALSE}). Logical stating whether plot should be saved into working directory.}

\item{dots}{(\code{default=FALSE}). If TRUE, progress dots are printed.}
}
\value{
Two plots one each for COVRATIO and PRESS statistics, giving the influence of each block on precision of the parameter estimates and the sensitivity of model predictions.
List object:
\item{influenceData}{List of \code{blocks}, COVRATIO statistics and PRESS statistics used for making the plot of PRESS and COVRATIO statistics.}
\item{influencePoints}{Row id of blocks in \code{influenceData} that lie outside the 95\% quantile of COVRATIO statistics and above the 95\% quantile of PRESS statistics.}
}
\description{
Assessing the influece of each correlated block on both the precision of the parameter estimates (COVRATIO statistics) and the sensitivity of model predictions (PRESS statistics).
}
\details{
Always run \code{\link{timeInfluenceCheck}} first to see how long it will take to produce the plots.
}
\examples{
# load data
data(ns.data.re)

ns.data.re$blockid<-paste(ns.data.re$GridCode, ns.data.re$Year, ns.data.re$MonthOfYear, 
                    ns.data.re$DayOfMonth, sep='')
ns.data.re$blockid<-as.factor(ns.data.re$blockid)

model<-gamMRSea(birds ~ observationhour + as.factor(floodebb) + as.factor(impact),  
              family='poisson', data=ns.data.re)

timeInfluenceCheck(model, ns.data.re$blockid)

\dontrun{
# **WARNING** this example takes a long time
influences<-runInfluence(model, ns.data.re$blockid)
}
}
