% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runPartialPlots.R
\name{runPartialPlots}
\alias{runPartialPlots}
\title{Plot partial plots for each of the variables listed in \code{factorlist.in} or \code{varlist.in}.}
\usage{
runPartialPlots(model, data, factorlist.in = NULL, varlist.in = NULL,
  showKnots = FALSE, type = "response", partial.resid = FALSE,
  save = FALSE, savedata = F, label = NULL)
}
\arguments{
\item{model}{Fitted model object (glm or gam)}

\item{data}{Data frame of data information used to fit \code{model}}

\item{factorlist.in}{(\code{default=NULL}). Vector or names of factor variables}

\item{varlist.in}{(\code{default=NULL}). Vector of names of continuous variables}

\item{showKnots}{(\code{default=FALSE}). Logical stating whether knot locations should be plotted.}

\item{type}{(\code{default='responss'}).  Character stating whether to return partial plots on the scale of the link function or the response.}

\item{partial.resid}{(\code{default=FALSE}).  Logical stating whether to include partial residuals on the plot.}

\item{save}{(\code{default=FALSE}). Logical stating whether plot should be saved into working directory.}

\item{savedata}{(\code{default=FALSE}). Logical stating whether the data used to make the plots should be saved into the working directory.  The object is called PartialData_'variablename'.RData}

\item{label}{(\code{default=NULL}).  This enables the user to specify a character label for the plots saved to the working directory. This may also be used to specify an alternative directory.}
}
\value{
Partial plots, one for each covariate in \code{factorlist.in} and \code{varlist.in}
}
\description{
Plot partial plots for each of the variables listed in \code{factorlist.in} or \code{varlist.in}.
}
\examples{

#' # load data
data(ns.data.re)

model<-glm(birds ~ observationhour + as.factor(floodebb) + as.factor(impact),
           family='quasipoisson', data=ns.data.re)

runPartialPlots(model, ns.data.re, factorlist.in=c('floodebb', 'impact'),
                varlist.in=c('observationhour'))
runPartialPlots(model, ns.data.re, factorlist.in=c('floodebb', 'impact'),
                varlist.in=c('observationhour'), type='link')
runPartialPlots(model, ns.data.re, factorlist.in=c('floodebb', 'impact'),
                varlist.in=c('observationhour'), partial.resid=TRUE)

}

