% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runSALSA1D.R
\name{runSALSA1D}
\alias{runSALSA1D}
\title{Running SALSA for continuous one-dimensional covariates.}
\usage{
runSALSA1D(
  initialModel,
  salsa1dlist,
  varlist,
  factorlist = NULL,
  predictionData = NULL,
  varlist_cyclicSplines = NULL,
  splineParams = NULL,
  datain,
  removal = FALSE,
  panelid = NULL,
  suppress.printout = FALSE,
  logfile = FALSE
)
}
\arguments{
\item{initialModel}{The best fitting \code{CReSS} model with no continuous covariates specified.  This must be a model of class \code{glm}.}

\item{salsa1dlist}{Vector of objects required for \code{runSALSA1D}: \code{fitnessMeasure}, \code{minKnots_1d}, \code{maxKnots_1d}, \code{startKnots_1d} \code{degree}, \code{maxIterations} \code{gap}.}

\item{varlist}{Vector of variable names for the covariates required for knot selection}

\item{factorlist}{vector of factor variables specified in \code{initialModel}.  Specified so that a check can be made that there are non-zero counts in all levels of each factor. Uses the function \code{checkfactorlevelcounts}. Default setting is NULL.}

\item{predictionData}{The data for which predictions are to be made. Column names must correspond to the data in \code{initialModel}. If predictionData is not specified (\code{NULL}), then the range of the data is used to create the smooth terms.}

\item{splineParams}{List object containing information for fitting splines to the covariates in \code{varlist}. If not specified (\code{NULL}) this object is created and returned. See \code{\link{makesplineParams}} for details.}

\item{datain}{Data used to fit the initial Model.}

\item{removal}{(Default: \code{FALSE}). Logical stating whether a selection procedure should be done to choose smooth, linear or removal of covariates.  If \code{FALSE} all covariates are returned and smooth. If \code{TRUE} then cross-validation is used to make model selection choices. The folds are specified by a column in the dataset called \code{foldid}.}

\item{panelid}{Vector denoting the panel identifier for each data point (if robust standard errors are to be calculated). Defaults to data order index if not given.}

\item{suppress.printout}{(Default: \code{FALSE}. Logical stating whether to show the analysis printout.}

\item{logfile}{(Default: \code{FALSE}). Logical stating whether to store a log file of the analysis printout.}
}
\value{
A list object is returned containing 4 elements:

\item{bestModel}{A model object of class \code{gam.MRSea} from the best model fitted}
\item{modelFits1D}{A list object with an element for each new term fitted to the model.  The first element is a model fitted with a knot at the mean for each of the covariates (startmodel) in \code{varlist}.  Within the first element, the current fit and formula of the start model.

The second element is the result of SALSA on the first term in \code{varlist}.  Within this element:
\itemize{
\item \code{term}: term of interest
\item \code{kept}: Statement of whether the term is kept in the model (yes- initial knots, yes - new knots, yes -linear or no)
\item \code{basemodelformula}: the resulting model formula.  If \code{kept=yes} or \code{kept=linear} then the term of interest is included in the model otherwise it is removed.
\item \code{knotSelected}: the knots chosen for the term of interest (NA if term removed or linear)
\item \code{baseModelFits}: fit statistics for the resulting formula
\item \code{modelfits}: fit statistics for the model with the term included (same as resulting formula if \code{kept=yes})
}

This continues till all covariates in \code{varlist} have been through SALSA.}
\item{fitstat}{The final fit statistic of \code{bestModel}.  The type of statistic was specified in \code{salsa1dlist}.}
\item{keptvarlist}{The covariates from \code{varlist} that have been retained in the model}
}
\description{
This function finds spatially adaptive knot locations for one or more continuous one-dimensional covariates.
}
\details{
There must be columns called \code{response} (response variable) and \code{foldid} (for cross-validation calculation) in the data used in the initial model to be fitted. If the data is proportion, then there should be two columns called \code{successess} and \code{failures}.

The object \code{salsa1dlist} contains parameters for the \code{runSALSA1D} function.

\code{fitnessMeasure}. The criterion for selecting the `best' model.  Available options: AIC, AIC_c, BIC, QIC_b, QAIC, QBIC, tweedie_AIC, tweedie_BIC, cv.gamMRSea (use fit.opts in salsa1dlist to specify seed, folds, cost function (Defaults: \code{fit.opts=list(cv.gamMRSea.seed=357, K=10, cost=function(y, yhat) mean((y - yhat)^2))}) or N for the BIC penalty (Default: \code{fit.opts$N = nrow(data)})).

\code{minKnots_1d}. Minimum number of knots to be tried.

\code{maxKnots_1d}. Maximum number of knots to be tried.

\code{startKnots_1d}. Starting number of knots (spaced at quantiles of the data).

\code{degree}. The degree of the B-spline. Does not need to be specified if \code{splineParams} is a parameter in \code{runSALSA1D}.

\code{maxIterations}.The exchange/improve steps will terminate after maxIterations if still running.

\code{gaps}. The minimum gap between knots (in unit of measurement of explanatory), usually set to zero.

\code{splines}. Specify the spline basis for each term.  Choose one of "bs" (B-spline), "cc" (cyclic-cubic) or "ns" (natural spline).

\code{minKnots_1d}, \code{maxKnots_1d}, \code{startKnots_1d} and \code{gaps} are vectors the same length as \code{varlist}.  This enables differing values of these parameters for each covariate.

The initial model contains all the factor level covariates and any covariates of interest that are not specified in the \code{varlist} argument of \code{runSALSA1D}

\emph{Note:} The algorithm may remove variables in \code{varlist} but not the variables in \code{factorlist}.  If there is no better model than with a knot at the mean, the output will include that covariate with a knot at the mean.  The best model with a given smooth term is tested both against a model with the term as linear or removed. Cross-Validation is used in the selection process.
}
\examples{
# load data
data(ns.data.re)
# load prediction data
data(ns.predict.data.re)

varlist=c('DayOfMonth')


# set initial model without the spline terms in there 
# (so all other non-spline terms)
ns.data.re$response<- ns.data.re$birds
initialModel<- glm(response ~ 1 + offset(log(area)), 
                    family='quasipoisson',data=ns.data.re)

#set some input info for SALSA
salsa1dlist<-list(fitnessMeasure = 'QBIC', 
                  minKnots_1d=c(1), 
                  maxKnots_1d = c(3), 
                  startKnots_1d = c(1), 
                  degree=c(2),
                  gaps=c(0))

# run SALSA
salsa1dOutput<-runSALSA1D(initialModel = initialModel,
                         salsa1dlist = salsa1dlist,
                         varlist = varlist,
                         predictionData = ns.predict.data.re,
                         datain = ns.data.re)

}
\references{
Walker, C.; M. Mackenzie, C. Donovan and M. O'Sullivan. SALSA - a Spatially Adaptive Local Smoothing Algorithm. Journal of Statistical Computation and Simulation, 81(2):179-191, 2010
}
\author{
Lindesay Scott-Hayward, University of St Andrews; Cameron Walker, Department of Engineering Science, University of Auckland.
}
