% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateNoise.R
\name{rzibinom}
\alias{rzibinom}
\title{rzibinom function from the VGAM package}
\usage{
rzibinom(n, size, prob, pstr0 = 0)
}
\description{
rzibinom function from the VGAM package
}
\author{
VGAM package
}
