% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{selectFctrStartk}
\alias{selectFctrStartk}
\title{Select starting knots for multi-level factor interaction}
\usage{
selectFctrStartk(knots, nk, s.eed = 1)
}
\arguments{
\item{knots}{Coordinates of the knot locations in columns 1 and 2 and the associated factor level in the third.}

\item{nk}{The number of knots to be selected for each factor level.  This may be a single number (for all factor levels) or a vector of numbers to allow a different number of knots to be selected for each factor level.}

\item{s.eed}{\code{default = 1}. Set the seed for the selection to ensure reproducibility.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

This function finds a number of starting knot locations for different factor levels and is used for creating an interaction term.
}
\details{
The function returns the row number of the \code{knots} parameter for the selected knots.
}
\examples{

#' # load data
data(nysted.analysisdata)

myknots <- selectFctrKnots(nysted.analysisdata[,c('x.pos', 'y.pos', 'impact')], 
                           nk=150)
startknotlocs <- selectFctrStartk(myknots, nk=5, s.eed = 4)


}
