% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.gamMRSea}
\alias{summary.gamMRSea}
\title{Summarising model fits from models fitted using the MRSea package.}
\usage{
\method{summary}{gamMRSea}(object, dispersion = NULL, digits = 7, varshortnames = NULL, ...)
}
\arguments{
\item{object}{an object of class "gamMRSea", usually, a result of a call from the MRSea package.}

\item{dispersion}{the dispersion parameter for the family used. Either a single numerical value or NULL (the default), when it is inferred from object (see 'Details').}

\item{digits}{the number of significant digits to use when printing.}

\item{varshortnames}{vector stating the short versions of the covariate names if required.}

\item{...}{further arguments passed to or from other methods.}
}
\description{
(object, dispersion = NULL, varshortnames=NULL, ...)
}
\details{
\code{print.summary.gamMRSea} tries to be smart about formatting the coefficients, standard errors, etc. and additionally gives 'significance stars' if signif.stars is TRUE. The coefficients component of the result gives the estimated coefficients and their estimated standard errors (raw and robust), together with their ratio (from robust s.e.). The third column gives the robust standard errors calculated using the sandwich estimator.  If no correlation is present, the second and third columns are the same as the sandwich estimator is not used when data points are independent. The fourth column is labelled Wald and gives the Wald test statistic, based on the robust standard errors. The fifth column gives the two-tailed p-value corresponding to the Wald test ().

Aliased coefficients are omitted in the returned object but restored by the print method.

Correlations are printed to two decimal places (or symbolically): to see the actual correlations print summary(object)$correlation directly.

\code{summary.gamMRSea} returns an object of class "summary.gamMRSea", a list with components
call	the component from object.
family	the component from object.
deviance	the component from object.
contrasts the component from object.
df.residual	the component from object.
null.deviance	the component from object.
df.null	the component from object.
deviance.resid	the deviance residuals: see residuals.glm.
coefficients	the matrix of coefficients, standard errors, z-values and p-values. Aliased coefficients are omitted.
aliased	named logical vector showing if the original coefficients are aliased.
dispersion	either the supplied argument or the inferred/estimated dispersion if the latter is NULL.
df	a 3-vector of the rank of the model and the number of residual degrees of freedom, plus number of coefficients (including aliased ones).
cov.unscaled	the unscaled (dispersion = 1) estimated covariance matrix of the estimated coefficients.
cov.scaled	ditto, scaled by dispersion.
correlation	(only if correlation is true.) The estimated correlations of the estimated coefficients.
symbolic.cor	(only if correlation is true.) The value of the argument symbolic.cor.
}
\note{
Code adapted from \code{summary.glm}
}
\examples{

# load data
data(ns.data.re)
ns.data.re$foldid<-getCVids(ns.data.re, folds=5)
 
model<-gamMRSea(birds ~ observationhour + as.factor(floodebb) + as.factor(impact),  
              family='poisson', data=ns.data.re)
summary(model)

}
\author{
Lindesay Scott-Hayward, University of St Andrews.
}
