% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.m.R
\name{summary.m}
\alias{summary.m}
\title{Summary function for MRSea models to shorten names.}
\usage{
summary.m(model, varlist = NULL, facls = NULL)
}
\arguments{
\item{model}{Fitted model object (glm or geeglm)}

\item{varlist}{(\code{default=NULL}). Vector of names of continuous variables.}

\item{facls}{(\code{default=NULL}). Vector or names of factor variables, as written in the model formula. e.g. including 'as.factor' wrappers.}
}
\value{
summary object as for \code{summary.glm} but with the variable names shortened for ease of viewing.
}
\description{
Summary function for MRSea models to shorten names.
}
\examples{
# load data
data(ns.data.re)

splineParams<-makesplineParams(data=ns.data.re, varlist=c('observationhour'))

# make column with foldid for cross validation calculation
ns.data.re$foldid<-getCVids(ns.data.re, folds=5)

#' # set initial model without the spline terms in there 
# (so all other non-spline terms)
ns.data.re$response<- ns.data.re$birds
initialModel<- glm(response ~ as.factor(floodebb) + as.factor(impact) + offset(log(area)), family='quasipoisson',data=ns.data.re)

#set some input info for SALSA
salsa1dlist<-list(fitnessMeasure = 'QBIC', minKnots_1d=c(1), maxKnots_1d = c(5), startKnots_1d = c(2), degree=c(2), maxIterations = 10, gaps=c(0))

# run SALSA
salsa1dOutput<-runSALSA1D_withremoval(initialModel, salsa1dlist, varlist=c('observationhour'), factorlist=c('floodebb', 'impact'), ns.predict.data.re, splineParams=splineParams, datain=ns.data.re)

summary.m(salsa1dOutput$bestModel,varlist=c('observationhour'), facls=c('as.factor(floodebb)', 'as.factor(impact)'))

# compare output with:
summary(salsa1dOutput$bestModel)

}

