% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaryshortnames.R
\name{summaryshortnames}
\alias{summaryshortnames}
\title{Shortening names in summary object}
\usage{
summaryshortnames(object, varshortnames)
}
\arguments{
\item{object}{a model object}

\item{varshortnames}{a list of short variable names}
}
\description{
Shortening names in summary object
}

