% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update.gammrsea.R
\name{update.gamMRSea}
\alias{update.gamMRSea}
\title{Update function for gamMRSea model object}
\usage{
\method{update}{gamMRSea}(object, formula., ..., evaluate = TRUE, panels = NULL)
}
\description{
Update function for gamMRSea model object
}
\note{
Code adapted from \code{update.glm}
}
\author{
Lindesay Scott-Hayward, University of St Andrews.
}
