% Generated by roxygen2 (4.0.2): do not edit by hand
\name{choose.radii}
\alias{choose.radii}
\title{Function to choose the radii for the CReSS local radial basis function}
\usage{
choose.radii(currentFit, indices, radiusIndices, radii, out.lm, dists, aR,
  baseModel, fitnessMeasure, response, models, interactionTerm, data, initDisp)
}
\description{
Function to choose the radii for the CReSS local radial basis function
}

