% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exchange.step_2d.R
\name{exchange.step_2d}
\alias{exchange.step_2d}
\title{Function for exchanging knot locations and re-fitting model to find best one}
\usage{
exchange.step_2d(
  gap,
  knotDist,
  radii,
  dists,
  explData,
  response,
  knotgrid,
  maxIterations,
  fitnessMeasure,
  point,
  knotPoint,
  position,
  aR,
  BIC,
  track,
  out.lm,
  improveEx,
  maxKnots,
  tol = 0,
  baseModel,
  radiusIndices,
  models,
  interactionTerm,
  data,
  initDisp,
  cv.opts,
  basis,
  printout
)
}
\description{
Function for exchanging knot locations and re-fitting model to find best one
}
\author{
Cameron Walker, Department of Enginering Science, University of Auckland and Lindesay Scott-Hayward (University of St Andrews)
}
