% Generated by roxygen2 (4.0.2): do not edit by hand
\name{fit.thinPlate_2d}
\alias{fit.thinPlate_2d}
\title{Function to fit a local radial basis function (CReSS) as a two dimensional smooth}
\usage{
fit.thinPlate_2d(fitnessMeasure, dists, aR, radii, baseModel, radiusIndices,
  models, currentFit, interactionTerm, data, initDisp)
}
\description{
Function to fit a local radial basis function (CReSS) as a two dimensional smooth
}

