% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getCV_CReSS}
\alias{getCV_CReSS}
\title{Calculate cross-validation score for a CReSS type model}
\usage{
getCV_CReSS(datain, baseModel, splineParams = NULL, vector = FALSE)
}
\arguments{
\item{baseModel}{glm or CReSS type model object}

\item{splineParams}{list object containing information for fitting one and two dimensional splines. See \code{\link{makesplineParams}} for more details.}

\item{vector}{Logical indicating whether the vector of scores is returned (TRUE) or if the mean score is returned (FALSE).}

\item{data}{Data frame containing columns of covariates contained in \code{baseModel}.}
}
\description{
Calculate cross-validation score for a CReSS type model
}
\details{
There must be a column in the data called \code{foldid}, which can be created using \code{\link{getCVids}}.  This column defines the folds of data for the CV calculation.  If this is not provided, by default random allocation is given to 5 folds.

The cost function for this CV is a mean squared error.
}
\examples{
# load data
data(ns.data.re)
# load prediction data
data(ns.predict.data.re)

splineParams<-makesplineParams(data=ns.data.re, varlist=c('observationhour', 'DayOfMonth'))
# set some input info for SALSA
ns.data.re$response<- ns.data.re$birds
salsa1dlist<-list(fitnessMeasure = 'QICb', minKnots_1d=c(2,2), maxKnots_1d = c(20, 20),
               startKnots_1d = c(2,2), degree=c(2,2), maxIterations = 10, gaps=c(1,1))

# set initial model without the spline terms in there
# (so all other non-spline terms)
initialModel<- glm(response ~ as.factor(floodebb) + as.factor(impact) + offset(log(area)),
                  family='quasipoisson',data=ns.data.re)

# run SALSA
salsa1dOutput<-runSALSA1D(initialModel, salsa1dlist, varlist=c('observationhour','DayOfMonth'),
            factorlist=c('floodebb', 'impact'), ns.predict.data.re, splineParams=splineParams)

# make blocking structure and fold structure
ns.data.re$blockid<-paste(ns.data.re$GridCode, ns.data.re$Year, ns.data.re$MonthOfYear,
                    ns.data.re$DayOfMonth, sep='')
ns.data.re$blockid<-as.factor(ns.data.re$blockid)
ns.data.re$foldid<-getCVids(ns.data.re, folds=5, block='blockid')

# calculate CV
cv1<-getCV_CReSS(ns.data.re, salsa1dOutput$bestModel, salsa1dOutput$splineParams)
}

