% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCVfoldID.R
\name{getCVids}
\alias{getCVids}
\title{IDs for running cross validation}
\usage{
getCVids(data, folds, block = NULL)
}
\arguments{
\item{data}{data used in regression model}

\item{folds}{integer number of validation data sets}

\item{block}{column in data indicating the blocking structure for cross-validation (if \code{block} = NULL, individual observations will be used as blocks)}
}
\description{
This function creates a string of integers which will be used for pointing to the right subsets of data for cross validation of regression objects
}
\details{
The function returns a random sequence of 1:folds of the same length as observations in data. It is called by other functions, e.g. \code{\link{getCV_CReSS}}.
}
\examples{
# load data
data(ns.data.re)

CVids<-getCVids(ns.data.re, 5)


}

