% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/improve.step_2d.R
\name{improve.step_2d}
\alias{improve.step_2d}
\title{Function to move knots to neighbours to see if there is any improvement in fit}
\usage{
improve.step_2d(gap, knotDist, radii, dists, explData, num, response, knotgrid,
  maxIterations, fitnessMeasure, point, knotPoint, position, aR, BIC, track,
  out.lm, improveNudge, tol = 0, baseModel, radiusIndices, models,
  interactionTerm, data, initDisp, seed.in, basis)
}
\description{
Function to move knots to neighbours to see if there is any improvement in fit
}
\author{
Cameron Walker, Department of Enginering Science, University of Auckland.
}

