% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MRSea-package.r
\docType{_PACKAGE}
\name{MRSea}
\alias{MRSea}
\title{MRSea}
\description{
Examines animal survey data for signs of changes in animal abundance and distribution following marine renewables development. The functions of this package can be used to analyse many types of data including segmented line transect data and nearshore vantage point data.  The package fits Generalised Additive Models (GAMs) to one and two dimensional splines/bases utilising a spatially adaptive framework for knot number and location. In one dimension, B-splines, cyclic cubics and natural cubics are available and in two dimensions, the radial bases can be Gaussian or exponential and use Euclidean or Geodesic (or other user specified) distances. Non-parametric bootstrapping is available to estimate uncertainty and if residual correlation is present, robust standard errors maybe used (given a panel structure). Several model assessment tools are also available.
}
