% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MRSea-package.r
\docType{package}
\name{MRSea}
\alias{MRSea}
\title{MRSea}
\description{
Examines animal survey data for signs of changes in animal abundance and distribution following marine renewables development. The functions of this package can be used to analyse segmented line transect data and nearshore vantage point data. The package includes functions for fitting spatially adaptive one and 2D smoothers using SALSA and CReSS. Euclidean or Geodesic distancees can be used to underpin the smoothed 2D surface and a choice of Gaussian or  exponential radial basis functions are available.  Non-parametric bootstrapping is available to estimate uncertainty. Several model assessment tools are also available. Recent updates include the direct estimation of robust standard errors, given a panel structure.
}
