% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/runACF.R
\name{plotacf}
\alias{plotacf}
\title{run functions to create acf matrix and plot the results}
\usage{
plotacf(acfmat)
}
\arguments{
\item{acfmat}{Matrix of output from \code{acffunc} (blocks x max block length).}
}
\description{
run functions to create acf matrix and plot the results
}

