% Generated by roxygen2 (4.0.2): do not edit by hand
\name{return.reg.spline.fit.2d}
\alias{return.reg.spline.fit.2d}
\title{Wrapper function for running SALSA2D}
\usage{
return.reg.spline.fit.2d(splineParams, startKnots, winHalfWidth,
  fitnessMeasure = "BIC", maxIterations = 10, tol = 0, baseModel = NULL,
  radiusIndices = NULL, initialise = TRUE, initialKnots = NULL,
  interactionTerm = NULL, knot.seed = 10)
}
\description{
Wrapper function for running SALSA2D
}
\author{
Cameron Walker, Department of Enginering Science, University of Auckland.
}

