% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runACF.R
\name{runACF}
\alias{runACF}
\title{run functions to create acf matrix and plot the results}
\usage{
runACF(block, model, store = FALSE, save = F, suppress.printout = FALSE)
}
\arguments{
\item{block}{Vector of blocks that identify data points that are correlated}

\item{model}{Fitted model object (glm or gam)}

\item{store}{(\code{default=F}). Logical stating whether a list of the matrix of correlations is stored (output from \code{acffunc}.)}

\item{save}{(\code{default=FALSE}). Logical stating whether plot should be saved into working directory.}
}
\value{
Plot of lag vs correlation.  Each grey line is the correlation for each individual block in \code{block}.  The red line is the mean values for each lag.

If \code{store=TRUE} then the matrix of correlations (nblocks x length_max_block) is returned and \code{plotacf} may be used to plot the acf.
}
\description{
run functions to create acf matrix and plot the results
}
\examples{
# load data
data(ns.data.re)

model<-glm(birds ~ observationhour + as.factor(floodebb) + as.factor(impact), 
           family='quasipoisson', data=ns.data.re)

ns.data.re$blockid<-paste(ns.data.re$GridCode, ns.data.re$Year, ns.data.re$MonthOfYear, 
                    ns.data.re$DayOfMonth, sep='')
ns.data.re$blockid<-as.factor(ns.data.re$blockid)

runACF(ns.data.re$blockid, model)

}

