% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runSALSA.R
\name{runSALSA1D}
\alias{runSALSA1D}
\title{Running SALSA for continuous one-dimensional covariates.}
\usage{
runSALSA1D(initialModel, salsa1dlist, varlist, factorlist = NULL,
  predictionData, varlist_cyclicSplines = NULL, splineParams = NULL)
}
\arguments{
\item{initialModel}{The best fitting \code{CReSS} model with no continuous covariates specified}

\item{salsa1dlist}{Vector of objects required for \code{runSALSA1D}: \code{fitnessMeasure}, \code{minKnots_1d}, \code{maxKnots_1d}, \code{startKnots_1d} \code{degree}, \code{maxIterations} \code{gap}.}

\item{varlist}{Vector of variable names for the covariates required for knot selection}

\item{factorlist}{vector of factor variables specified in \code{initialModel}.  Specified so that a check can be made that there are non-zero counts in all levels of each factor. Uses the function \code{checkfactorlevelcounts}. Default setting is NULL.}

\item{predictionData}{The data to be predicted for. column names correspond to the data in \code{initialModel}}

\item{varlist_cyclicSplines}{Vector of variable names for covariates to be modelled with cyclic cubic splines.  This must be a subset of \code{varlist}.The default is \code{NULL}}

\item{splineParams}{List object containing information for fitting splines to the covariates in \code{varlist}. If not specified (\code{NULL}) this object is created and returned. See \code{\link{makesplineParams}} for details.}
}
\value{
A list object is returned containing 4 elements:

\item{bestModel}{A glm model object from the best model fitted}
\item{modelFits1D}{A list object with an element for each new term fitted to the model.  The first element is a model fitted with a knot at the mean for each of the covariates in \code{varlist}.  Within the first element, the model term of interest, the current fit, knots and formula.  The second element is the result of SALSA on the first term in \code{varlist}.  Within this element, the knots chosen and the improvement in model fit are presented \code{$modelfits}.  This continues till all covariates in \code{varlist} have been through SALSA.}
\item{splineParams}{The updated spline parameter object, with the new (if chosen) knot locations for each covariate in \code{varlist}}
\item{fitstat}{The final fit statistic of \code{bestModel}.  The type of statistic was specified in \code{salsa1dlist}.}
}
\description{
This function finds spatially adaptive knot locations for one or more continuous one-dimensional covariates.
}
\details{
There must be a column called \code{response} in the data, which is the response variable used in the initial model to be fitted.

The object \code{salsa1dlist} contains parameters for the \code{runSALSA1D} function.
  
\code{fitnessMeasure}. The criterion for selecting the `best' model.  Available options: AIC, AIC_c, BIC, QIC_b.

   \code{minKnots_1d}. Minimum number of knots to be tried.
   
   \code{maxKnots_1d}. Maximum number of knots to be tried.
   
   \code{startKnots_1d}. Starting number of knots (spaced at quantiles of the data).
   
   \code{degree}. The degree of the B-spline. Does not need to be specified if \code{splineParams} is a parameter in \code{runSALSA1D}.
   
   \code{maxIterations}.The exchange/improve steps will terminate after maxIterations if still running.
   
   \code{gaps}. The minimum gap between knots (in unit of measurement of explanatory).
   
   
\code{minKnots_1d}, \code{maxKnots_1d}, \code{startKnots_1d} and \code{gaps} are vectors the same length as \code{varlist}.  This enables differing values of these parameters for each covariate.

 The initial model contains all the factor level covariates and any covariates of interest that are not specified in the \code{varlist} argument of \code{runSALSA1D} 

\emph{Note:} The algorithm will not remove variables in \code{varlist}.  If there is no better model than with a knot at the mean, the output will include that covariate with a knot at the mean.  The user must decide if the covariate is required in the model as a linear term instead.
}
\examples{
# load data
data(ns.data.re)
# load prediction data
data(ns.predict.data.re)

splineParams<-makesplineParams(data=ns.data.re, varlist=c('observationhour', 'DayOfMonth'))
#set some input info for SALSA
ns.data.re$response<- ns.data.re$birds

#' # set initial model without the spline terms in there 
# (so all other non-spline terms)
initialModel<- glm(response ~ as.factor(floodebb) + as.factor(impact) + offset(log(area)), 
                    family='quasipoisson',data=ns.data.re)

salsa1dlist<-list(fitnessMeasure = 'QICb', minKnots_1d=c(2,2), maxKnots_1d = c(20, 20), 
                  startKnots_1d = c(2,2), degree=c(2,2), maxIterations = 10, gaps=c(1,1))
# run SALSA
salsa1dOutput<-runSALSA1D(initialModel, salsa1dlist, varlist=c('observationhour', 'DayOfMonth'), 
                 factorlist=c('floodebb', 'impact'), ns.predict.data.re, splineParams=splineParams)

}
\references{
Walker, C.; M. Mackenzie, C. Donovan and M. O'Sullivan. SALSA - a Spatially Adaptive Local Smoothing Algorithm. Journal of Statistical Computation and Simulation, 81(2):179-191, 2010
}

