% Generated by roxygen2 (4.0.2): do not edit by hand
\name{runSALSA2D}
\alias{runSALSA2D}
\title{Running SALSA for a spatial smooth with a CReSS basis}
\usage{
runSALSA2D(model, salsa2dlist, d2k, k2k, splineParams = NULL, tol = 0,
  chooserad = F)
}
\arguments{
\item{model}{A model with no spatial smooth}

\item{salsa2dlist}{Vector of objects required for \code{runSALSA2D}: \code{fitnessMeasure}, \code{knotgrid}, \code{startKnots}, \code{minKnots}, code{maxKnots}, \code{r_seq}, \code{gap}, \code{interactionTerm}.}

\item{d2k}{(n x k) Matrix of distances between all data points in \code{model} and all valid knot locations specified in \code{knotgrid}
#'}

\item{k2k}{(k x k) Matrix of distances between all valid knot locations specified in \code{knotgrid}}

\item{splineParams}{(default \code{=NULL}) List object containng output from runSALSA (e.g. knot locations for continuous covariates)}
}
\value{
The spline paramater object that is return now contains a list in the first element (previously reserved for the spatial component).  This list contains the objects required for the SALSA2D fitting process:

\item{knotDist}{Matrix of knot to knot distances (k x k).  May be Euclidean or geodesic distances. Must be square and the same dimensions as \code{nrows(na.omit(knotgrid))}.  Created using \code{\link{makeDists}}.}
\item{radii}{Sequence of range parameters for the CReSS basis from local (small) to global (large).  Determines the range of the influence of each knot.}
\item{dist}{ Matrix of distances between data locations and knot locations (n x k). May be Euclidean or geodesic distances. Euclidean distances created using \code{\link{makeDists}}.}
\item{gridresp}{The first column of knotgrid.}
\item{grid}{Index of knotgrid locations.  Should be same length as \code{knotgrid} but with x=integer values from 1 to number of unique x-locations and y= integer values from 1 to number of unique y-locations.}
\item{datacoords}{Coordinates of the data locations}
\item{response}{Vector of response data for the modelling process}
\item{knotgrid}{Grid of legal knot locations.  Must be a regular grid with c(NA, NA) for rows with an illegal knot.}
\item{minKnots}{Minimum number of knots to be tried.}
\item{maxKnots}{Maximum number of knots to be tried.}
\item{gap}{Minimum gap between knots (in unit of measurement of \code{datacoords})}
\item{radiusIndices}{Vector of length startKnots identifying which radii (\code{splineParams[[1]]$radii}) will be used for each knot location (\code{splineParams[[1]]$knotPos})}
\item{knotPos}{Index of knot locations. The index identifies which knots (i.e. which rows) from \code{knotgrid} were selected by SALSA}
\item{invInd}{This is a vector of length the number of rows of \code{knotgrid}.  It is used to translate between \code{knotgrid} (used in SALSA) and \code{na.omit(knotgrid)} (used in \code{dist} and \code{LocalRadialFunction}).}
}
\description{
This function fits a spatially adaptive two dimensional smooth of spatial coordinates with knot number and location selected by SALSA.
}
\details{
There must be a column called \code{response} in the data, which is the response variable used in the initial model to be fitted.

The object \code{salsa2dlist} contains parameters for the \code{runSALSA2D} function.

   \code{fitnessMeasure}. The criterion for selecting the `best' model.  Available options: AIC, AIC_c, BIC, QIC_b.

   \code{knotgrid}. A grid of legal knot locations.  Must be a regular grid with \code{c(NA, NA)} for rows with an illegal knot.  An illegal knot position may be outside the study region or on land for a marine species for example.

   \code{knotdim}. The dimensions of the knot grid as a vector. (x, y)

   \code{startknots}. Starting number of knots (initialised as spaced filled locations).

   \code{minKnots}. Minimum number of knots to be tried.

   \code{maxKnots}. Maximum number of knots to be tried.

   \code{r_seq}. Sequence of range parameters for the CReSS basis from local (small) to global (large).  Determines the range of the influence of each knot. Sequence made using \code{\link{getRadiiChoices}}.

   \code{gap}. The minimum gap between knots (in unit of measurement of coordinates).
   \code{interactionTerm}. Specifies which term in \code{baseModel} the spatial smooth will interact with.  If \code{NULL} no interaction term is fitted.
}
\examples{
# load data
data(ns.data.re)
# load prediction data
data(ns.predict.data.re)
# load knot grid data
data(knotgrid.ns)

splineParams<-makesplineParams(data=ns.data.re, varlist=c('observationhour'))

#set some input info for SALSA
ns.data.re$response<- ns.data.re$birds

# make distance matrices for datatoknots and knottoknots
distMats<-makeDists(cbind(ns.data.re$x.pos, ns.data.re$y.pos), na.omit(knotgrid.ns))

# choose sequence of radii
r_seq<-getRadiiChoices(8, distMats$dataDist)

# set initial model without the spatial term
# (so all other non-spline terms)
initialModel<- glm(response ~ as.factor(floodebb) + as.factor(impact) + offset(log(area)),
                   family='quasipoisson', data=ns.data.re)

# make parameter set for running salsa2d
# I have chosen a gap parameter of 1000 (in metres) to speed up the process.
# Note that this means there cannot be two knots within 1000m of each other.

salsa2dlist<-list(fitnessMeasure = 'QICb', knotgrid = knotgrid.ns, knotdim = c(7, 9),
                  startKnots=6, minKnots=4, maxKnots=20, r_seq=r_seq, gap=1000,
                  interactionTerm="as.factor(impact)")

salsa2dOutput_k6<-runSALSA2D(initialModel, salsa2dlist, d2k=distMats$dataDist,
                            k2k=distMats$knotDist, splineParams=splineParams)
}
\author{
Cameron Walker
}
\references{
Scott-Hayward, L.; M. Mackenzie, C.Donovan, C.Walker and E.Ashe.  Complex Region Spatial Smoother (CReSS). Journal of computational and Graphical Statistics. 2013. doi: 10.1080/10618600.2012.762920

Scott-Hayward, L.. Novel Methods for species distribution mapping including spatial models in complex regions: Chapter 5 for SALSA2D methods. PhD Thesis, University of St Andrews. 2013
}

