% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runsTest.R
\name{runsTest}
\alias{runsTest}
\title{Runs Test for Randomness}
\usage{
runsTest(y, plot.it = FALSE, alternative = c("two.sided",
  "positive.correlated", "negative.correlated"), emp.distribution = NULL)
}
\description{
Runs Test for Randomness
}

