% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thinModels.R
\name{thinModels}
\alias{thinModels}
\title{function to thin the number of models}
\usage{
thinModels(models)
}
\description{
function to thin the number of models
}
\author{
Cameron Walker, Department of Enginering Science, University of Auckland.
}

